refactor: util함수 관심사별 분리 및 index.ts에서 export하도록 수정
