+++ src/utils/check/isUndefinedOrNull.ts
@@ -1,0 +1,4 @@
+export const isUndefinedOrNull = (
+  element: unknown
+): element is null | undefined =>
+  typeof element === 'undefined' || element === null;
