+++ src/utils/index.ts
@@ -1,4 +1,5 @@
-export const isUndefinedOrNull = (
-  element: unknown
-): element is null | undefined =>
-  typeof element === 'undefined' || element === null;
+export * from './check/isUndefinedOrNull';
+export * from './element/appendChildElement';
+export * from './element/getElement';
+export * from './input/getInputValue';
+export * from './input/initInputValue';
